Revert "Adding tab menu:"
