--- app/src/main/java/com/example/pousadaviladascores/Activities/Tab3ItensDeApartamentos.java
@@ -1,19 +1,0 @@
-package com.example.pousadaviladascores.Activities;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.fragment.app.Fragment;
-
-import com.example.pousadaviladascores.R;
-
-public class Tab3ItensDeApartamentos extends Fragment {
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
-    {
-        return inflater.inflate(R.layout.tab3_fragment_itens_de_apartamentos, container, false);
-    }
-}
