--- app/src/main/res/layout/tab2_fragment_apartamentos.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <TextView
-        android:id="@+id/textTab"
-        android:layout_width="369dp"
-        android:layout_height="54dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="27dp"
-        android:layout_marginLeft="27dp"
-        android:layout_marginTop="0dp"
-        android:layout_marginEnd="15dp"
-        android:text="@string/tab_text_2"
-        android:textSize="30sp"
-        android:layout_alignParentRight="true"
-        android:layout_marginRight="15dp" />
-
-</RelativeLayout>
